+++ CHANGELOG.md
@@ -35,6 +35,7 @@
     - can also parse Geoserver responses from DescribeFeatureType-Requests
     - fixed language pack in nested i18next json
     - correct handling of boolean attributes
+- Issue #863: Print: Consider offsets for point features with icon
 - Issue #1024: LayerInformation: Fix URL generation for given query parameters in service URL
 - Issue #1045: WFSSearch: Fields will be enabled correctly for multiple search instances now
 - Issue #1057: Filter: Checkbox Snippet does not response right in activation and deactivation.